+++ app/(pages)/layout.tsx
@@ -4,9 +4,10 @@ import UpdateNotice from '@/app/ui/UpdateNotice'
 import type { Metadata, Viewport } from 'next'
 import './globals.css'
 import { readLatestChanges } from '../lib/changes'
+import { env } from '@/app/lib/env'
 
 export const metadata: Metadata = {
-    metadataBase: 'http://localhost:3000',
+    metadataBase: env.BASE_URL,
     title: "Cal's cals",
     description: 'Locally produced calendars at Chalmers!',
     applicationName: "Cal's cals",
@@ -60,13 +61,7 @@ export default async function RootLayout({
                             <p>
                                 A better calendar experience at Chalmers,
                                 created and maintained by&nbsp;
-                                <a
-                                    href="https://wiki.chalmers.it/Cal"
-                                    target="_blank"
-                                >
-                                    Cal
-                                </a>
-                                .
+                                <a href="https://wiki.chalmers.it/Cal">Cal</a>.
                             </p>
                         </section>