+++ app/lib/adapter/TimeEditAdapter.ts
@@ -1,26 +1,41 @@
 import { Calendar, CalendarEvent } from 'iamcal'
-import { NextRequest } from 'next/server'
-import Adapter from './Adapter'
+import Adapter, { AdapterContext } from './Adapter'
 import HashSlicer from '../slicer/HashSlicer'
 import Slicer, { EventGroup, applySlicer } from '../slicer/Slicer'
 import {
-    AvailableGroup,
     createCalendarDescription,
     createCalendarName,
     createEventDescription,
     createEventLocation,
     createEventSummary,
     createExamEvents,
-    groupByOptions,
+    timeEditGroupByOptions,
     isGlobalEvent,
+    isGuCourseCode,
     parseEventData,
+    serializeEventData,
     shortenCourseCode,
-    TimeEditUrlExtras,
+    TimeEditGroupByOption,
+    createBookingEventSummary,
 } from '../timeedit'
+import {
+    AvailableGroup,
+    GroupedUrlExtras,
+    parseAllowedValues,
+    parseGroupBy,
+    prepareSetForComparison,
+} from '../group'
 
 export default class TimeEditAdapter extends Adapter {
     override createUrl(id: string): URL {
         const [category, filename] = id.split('.')
+
+        if (category === 'student-bookings') {
+            return new URL(
+                `https://cloud.timeedit.net/chalmers/web/student/my.ics?i=${filename}`
+            )
+        }
+
         return new URL(
             `https://cloud.timeedit.net/chalmers/web/${category}/${filename}.ics`
         )
@@ -28,7 +43,7 @@ export default class TimeEditAdapter extends Adapter {
 
     override getId(url: URL): string {
         const urlPattern =
-            /^(https|webcal):\/\/cloud\.timeedit\.net\/\w+\/web\/\w+\/[^/]+\.ics$/
+            /^(https|webcal):\/\/cloud\.timeedit\.net\/\w+\/web\/\w+\/[^/]+\.ics(\?i=\w+)?$/
         if (!urlPattern.test(url.href)) {
             throw new Error('Invalid URL')
         }
@@ -44,13 +59,26 @@ export default class TimeEditAdapter extends Adapter {
         const categoryPattern = /(?<=web\/).+?(?=\/)/
         const category = categoryPattern.exec(url.href)?.[0]
 
-        if (category !== 'public')
+        if (category !== 'public' && category !== 'student') {
             throw new Error(
-                'Unsupported category. Calendar must be from the "public" schedule ("Öppen schemavisning").'
+                'Unsupported category. Calendar must be from the "public" schedule ("Öppen schemavisning") or room bookings.'
             )
+        }
 
         const filenamePattern = /[^/]+?(?=\.ics)/
         const filename = filenamePattern.exec(url.href)![0]
+
+        const isBookingCalendar = category === 'student' && filename === 'my'
+        if (isBookingCalendar) {
+            const iPattern = /(?<=\?i=)\w+$/
+            const i = iPattern.exec(url.href)?.[0]
+            if (!i) throw new Error('Invalid URL')
+            return `student-bookings.${i}`
+        } else if (category === 'student') {
+            throw new Error(
+                'Unsupported category. Calendar must be from the "public" schedule ("Öppen schemavisning") or room bookings.'
+            )
+        }
 
         const id = `${category}.${filename}`
         return id
@@ -58,18 +86,39 @@ export default class TimeEditAdapter extends Adapter {
 
     override async patchCalendar(
         calendar: Calendar,
-        req?: NextRequest
+        context?: AdapterContext
     ): Promise<Calendar> {
+        // Update calendar description
+        const calendarDescription = createCalendarDescription(
+            calendar.getCalendarDescription()
+        )
+        calendar.setCalendarDescription(calendarDescription)
+
+        // Handle bookings calendar
+        const category = context?.id ? context.id.split('.')[0] : null
+        if (category === 'student-bookings') {
+            // Update calendar metadata
+            calendar.setCalendarName('TimeEdit Bookings')
+            return calendar
+        }
+
+        // Public calendar
         const addExams =
-            req == undefined ||
+            context?.req == undefined ||
             !['1', 'true', 't'].includes(
-                req.nextUrl.searchParams.get('noExam') ?? '0'
+                context.req.nextUrl.searchParams.get('noExam') ?? '0'
             )
 
         const keepGlobalEvents =
-            req != undefined &&
+            context?.req != undefined &&
             ['1', 'true', 't'].includes(
-                req.nextUrl.searchParams.get('keepGlobal') ?? '0'
+                context.req.nextUrl.searchParams.get('keepGlobal') ?? '0'
+            )
+
+        const hideGu =
+            context?.req != undefined &&
+            ['1', 'true', 't'].includes(
+                context.req.nextUrl.searchParams.get('hideGu') ?? '0'
             )
 
         const courseCodeSets: string[][] = []
@@ -81,6 +130,17 @@ export default class TimeEditAdapter extends Adapter {
             }
 
             const eventData = parseEventData(event)
+
+            if (hideGu && eventData.kurskod) {
+                eventData.kurskod = eventData.kurskod.filter(
+                    code => !isGuCourseCode(code)
+                )
+                // Patch event
+                event.setSummary(serializeEventData(eventData))
+                event.setDescription('')
+                event.setLocation('')
+            }
+
             const courseCodes = eventData.kurskod?.map(shortenCourseCode)
             if (courseCodes === undefined) return
 
@@ -106,21 +166,20 @@ export default class TimeEditAdapter extends Adapter {
             calendar.addComponents(examEvents)
         }
 
-        // Update calendar metadata
+        // Update calendar name
         const calendarName = createCalendarName(groupedCourseCodes)
         calendar.setCalendarName(calendarName)
-        const calendarDescription = createCalendarDescription(
-            calendar.getCalendarDescription()
-        )
-        calendar.setCalendarDescription(calendarDescription)
 
         return calendar
     }
 
-    override convertCalendar(calendar: Calendar, req?: NextRequest): Calendar {
-        if (req?.nextUrl.searchParams.get('group')) {
-            const groupBy = parseGroupBy(req)
-            const allowedValues = parseAllowedValues(req)
+    override convertCalendar(
+        calendar: Calendar,
+        context?: AdapterContext
+    ): Calendar {
+        if (context?.req?.nextUrl.searchParams.get('group')) {
+            const groupBy = parseGroupBy(context.req)
+            const allowedValues = parseAllowedValues(context.req)
             const slicer = createGroupSlicer(groupBy, allowedValues)
 
             // Include only the group which has events with the included values
@@ -128,21 +187,25 @@ export default class TimeEditAdapter extends Adapter {
             applySlicer(calendar, slicer, mask)
         }
 
-        calendar.getEvents().forEach(event => convertEvent(event))
+        calendar.getEvents().forEach(event => convertEvent(event, context))
         return calendar
     }
 
-    override getExtras(calendar: Calendar): TimeEditUrlExtras {
-        const groups: AvailableGroup[] = groupByOptions.map(option => ({
-            property: option,
-            values: {},
-        }))
+    override getExtras(
+        calendar: Calendar
+    ): GroupedUrlExtras<TimeEditGroupByOption> {
+        const groups: AvailableGroup<TimeEditGroupByOption>[] =
+            timeEditGroupByOptions.map(option => ({
+                property: option,
+                propertyIndex: timeEditGroupByOptions.indexOf(option),
+                values: {},
+            }))
 
         calendar.getEvents().forEach(event => {
             const data = parseEventData(event)
             // For each groupable property name
-            for (let i = 0; i < groupByOptions.length; i++) {
-                const property = groupByOptions[i]
+            for (let i = 0; i < timeEditGroupByOptions.length; i++) {
+                const property = timeEditGroupByOptions[i]
 
                 if (data[property]) {
                     // Event has property
@@ -159,74 +222,13 @@ export default class TimeEditAdapter extends Adapter {
             }
         })
 
-        const extras: TimeEditUrlExtras = {
+        const extras: GroupedUrlExtras<TimeEditGroupByOption> = {
+            name: calendar.getCalendarName(),
             groups: groups,
         }
 
         return extras
     }
-}
-
-/**
- * Get the index of the property to group by.
- * @param req The request to parse.
- * @returns The index of the property in {@link groupByOptions}.
- * @throws {Error} If the 'group' query parameter is missing or invalid.
- */
-export function parseGroupBy(req: NextRequest): number {
-    const groupByString = req.nextUrl.searchParams.get('group')
-    if (groupByString === undefined)
-        throw new Error(
-            "Unable to find property to group by. Missing query parameter 'group'"
-        )
-
-    let groupBy: number
-    try {
-        groupBy = parseInt(String(groupByString))
-    } catch {
-        throw new Error(
-            "Invalid query parameter 'group', must be a positive integer"
-        )
-    }
-    if (groupBy < 0) {
-        throw new Error(
-            "Invalid query parameter 'group', must be a positive integer"
-        )
-    }
-
-    return groupBy
-}
-
-/**
- * Get the allowed values for grouping from the request.
- *
- * One element represents one allowed combination of TimeEdit property values.
- * @param req The request to parse.
- * @returns A set of allowed values.
- * @throws {Error} If the request has no group values.
- */
-export function parseAllowedValues(req: NextRequest): Set<string> {
-    const serializedValues = req.nextUrl.searchParams.get('gi')
-    if (serializedValues == undefined)
-        throw new Error(
-            "Unable to get group index. Missing query parameter 'gi'"
-        )
-
-    const values = String(serializedValues)
-        .replace(/[^a-z0-9_ -]/g, '')
-        .split(' ')
-        .map(value =>
-            value === '_'
-                ? value
-                : value
-                      .split('_')
-                      .filter(v => v !== '')
-                      .sort()
-                      .join('_')
-        )
-        .filter(v => v !== '')
-
-    return new Set(values)
 }
 
 /**
@@ -240,10 +242,10 @@ export function createGroupSlicer(
     groupBy: number,
     allowedValues: Set<string>
 ): Slicer<EventGroup> {
-    if (groupBy >= groupByOptions.length) {
+    if (groupBy >= timeEditGroupByOptions.length) {
         throw new Error(`Unknown group by option ${groupBy}.`)
     }
-    const property = groupByOptions[groupBy]
+    const property = timeEditGroupByOptions[groupBy]
 
     /**
      * This hash function returns 1 for events that have any of the allowed
@@ -264,36 +266,22 @@ export function createGroupSlicer(
 }
 
 /**
- * Prepare a value to be used in grouping comparisons.
- * @param value The value to simplify.
- * @returns A URL friendly simplified version of the value.
- */
-export function prepareForComparison(value: string): string {
-    return value.toLowerCase().replace(/[^a-z0-9]/g, '-')
-}
-
-/**
- * Prepare a value to be used in grouping comparisons.
- * @param value The set to simplify, represents one combination of property values.
- * @returns A URL friendly simplified version of the value.
- */
-export function prepareSetForComparison(value: string[]): string {
-    return value
-        .map(prepareForComparison)
-        .filter(v => v !== '')
-        .sort()
-        .join('_')
-}
-
-/**
  * Update a TimeEdit event to follow the adapter format.
  * @param event The event from TimeEdit.
+ * @param context The context of the request to get the calendar.
  */
-export function convertEvent(event: CalendarEvent): void {
+export function convertEvent(
+    event: CalendarEvent,
+    context?: AdapterContext
+): void {
     const eventData = parseEventData(event)
+    const category = context?.id?.split('.')[0]
 
     // Summary
-    const summary = createEventSummary(eventData, event)
+    const summary =
+        category === 'student-bookings'
+            ? createBookingEventSummary(eventData, event)
+            : createEventSummary(eventData, event)
     if (summary) {
         event.setSummary(summary)
     } else {