+++ app/lib/changes.ts
@@ -50,6 +50,22 @@ const replaceHeadings: Plugin<[]> = () => {
     }
 }
 
+const removeTopHeading: Plugin<[]> = () => {
+    return (tree: mdast.Root) => {
+        // Find second heading
+        const firstIndex = tree.children.findIndex(node => {
+            return node.type === 'heading'
+        })
+        const secondIndex = tree.children.findIndex((node, i) => {
+            return i > firstIndex && node.type === 'heading'
+        })
+        // Do nothing if not found
+        if (secondIndex === -1) return
+
+        tree.children = tree.children.slice(secondIndex)
+    }
+}
+
 export async function readLatestChanges(
     filename: string = 'CHANGELOG.md'
 ): Promise<Changes | null> {
@@ -80,4 +96,18 @@ export async function readLatestChanges(
         name,
         body,
     }
+}
+
+export async function readAllChanges(
+    filename: string = 'CHANGELOG.md'
+): Promise<ReactNode | null> {
+    'use cache'
+    const text = await fs.readFile(filename)
+    const file = (await unified()
+        .use(remarkParse, { fragment: true })
+        .use(removeTopHeading)
+        .use(remarkRehype)
+        .use(rehypeReact, production)
+        .process(text)) as VFile
+    return file.result as ReactNode
 }