+++ app/ui/PageFooter.tsx
@@ -1,4 +1,6 @@
-const currentVersion: string | undefined = process.env.NEXT_PUBLIC_WEB_VERSION
+import { env } from '@/app/lib/env'
+
+const currentVersion = env.NEXT_PUBLIC_WEB_VERSION
 const versionString = currentVersion
     ? currentVersion.match(/^[0-9]+\.[0-9]+/)
         ? `v${currentVersion}`
@@ -10,17 +12,11 @@ export default function PageFooter() {
         <footer>
             <span>
                 Made with <span className="pixelnheart"></span> by{' '}
-                <a href="https://wiki.chalmers.it/Cal" target="_blank">
-                    Cal
-                </a>
-                .
+                <a href="https://wiki.chalmers.it/Cal">Cal</a>.
             </span>
             <span>
                 Is something wrong or could be improved?&nbsp;
-                <a
-                    href="https://github.com/olillin/cals-cals/issues"
-                    target="_blank"
-                >
+                <a href="https://github.com/olillin/cals-cals/issues">
                     Leave an issue
                 </a>{' '}
                 on GitHub!
@@ -28,7 +24,7 @@ export default function PageFooter() {
             <span>
                 {versionString && `Cal's cals ${versionString}. `}
                 See the{' '}
-                <a href="https://github.com/olillin/cals-cals" target="_blank">
+                <a href="https://github.com/olillin/cals-cals">
                     source code
                 </a>{' '}
                 on GitHub.