+++ CHANGELOG.md
@@ -42,3 +42,27 @@ This is a smaller update containing some minor changes and patches:
   calendar
 - Fixed the calendar picker grid again
 - Fixed some checkbox labels not being clickable
+
+## v2.3 Canvas calendars and QoL!
+
+This update adds a new calendar adapter for Canvas, granted Canvas calendars are
+not nearly as bad as TimeEdit so the main feature is being able to group by
+assignment/event or course code. There are some improvements to the format
+however! For assignments a link to the Canvas page is added to the top of the
+description and all descriptions now allow rich formatting for calendar apps
+which support it.
+
+But that's not all, other changes for this update include:
+
+- The TimeEdit calendar adapter has been moved to [/timeedit](./timeedit) and
+  been made more obvious in the navigation.
+- Proper subscribe buttons have been added to all calendars for _Google
+  Calendar_, _Apple Calendar_ and _Outlook_.
+- Added an option to the TimeEdit adapter to hide GU course codes.
+- Support has been added for
+  [TimeEdit booking calendars](https://cloud.timeedit.net/chalmers/web/student/my.html)
+  in the TimeEdit adapter.
+- Added a helpful message when a calendar cannot be grouped.
+- Added a changelog page.
+- Fixed a bug causing grouping by the wrong property.
+- Fixed metadata for media previews pointing to localhost.